pyproject.toml
[tool.poetry]
name = "bookit"
version = "0.1.0"
description = "BookIt API – FastAPI + PostgreSQL"
authors = ["You <odapo138@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.115.0"
udicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = "^2.0.32"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.2"
pydantic = "^2.8.2"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
email-validator = "^2.2.0"
structlog = "^24.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.8"
